Display of multiple treatment types using API data
Mouth and Tooth data is now returned from an API route on the server.
Using placeholder data for now.

3D viewer is able to show the following treatments:
- Fillings
- Crowns
- Root Canals
- Extractions
- Implants
- Veneers
- Sealant

Currently the viewer does not show bridges as we have not decided on a
way to depict these.
Missing teeth are not distinct from extracted teeth yet.

Currently using the colour-coding system from the App team. This may
need to be modified later for better realism.